+++ frontend/components/chat/friend/FriendInfoDialog.tsx
