+++ frontend/pages/chat/index.tsx
@@ -7,7 +7,7 @@ import { Header } from 'components/common/Header';
 import { ChatroomListItem } from 'components/chat/ChatroomListItem';
 import { ChatroomCreateButton } from 'components/chat/ChatroomCreateButton';
 import { ChatroomJoinButton } from 'components/chat/ChatroomJoinButton';
-import { FriendBar } from 'components/chat/FriendBar';
+import { FriendSidebar } from 'components/chat/friend/FriendSidebar';
 import { Chatroom, Message } from 'types/chat';
 import { useQueryUser } from 'hooks/useQueryUser';
 import { Loading } from 'components/common/Loading';
@@ -208,7 +208,7 @@ const Chat = () => {
             borderBottom: '1px solid',
           }}
         >
-          <FriendBar />
+          <FriendSidebar />
         </Grid>
       </Grid>
     </>
